Inline admin mutations in editor controls
- Move story admin controls into the editor v2 client component
- Call Convex mutations directly and remove the server action shim
- Keep pending state while toggling publish status or deleting approvals
